adding data seed for profile, refactoring Home Context splitting fetching logic
