+++ app/seed/x-simulator.profiles.json
@@ -1,0 +1,36 @@
+[{
+  "_id": {
+    "$oid": "668136d3f45300016a3dfe3e"
+  },
+  "name": "Aleister Crowley",
+  "role": "Admin",
+  "profileImg": "https://pbs.twimg.com/profile_images/1308385514744098816/oDXuaci__400x400.jpg",
+  "country": "USA",
+  "followingCount": 4533,
+  "followerCount": 2654,
+  "tweetCount": 4795
+},
+{
+  "_id": {
+    "$oid": "668136d3f45300016a3dfe3f"
+  },
+  "name": "Eliphaz Levi",
+  "role": "User",
+  "profileImg": "https://pbs.twimg.com/profile_images/1308385514744098816/oDXuaci__400x400.jpg",
+  "country": "Australia",
+  "followingCount": 4089,
+  "followerCount": 1758,
+  "tweetCount": 8317
+},
+{
+  "_id": {
+    "$oid": "668136d3f45300016a3dfe40"
+  },
+  "name": "Amhenotep",
+  "role": "Admin",
+  "profileImg": "https://pbs.twimg.com/profile_images/1308385514744098816/oDXuaci__400x400.jpg",
+  "country": "Canada",
+  "followingCount": 4477,
+  "followerCount": 4789,
+  "tweetCount": 9853
+}]
